requirements.txt
langchain
newsapi-python
chromadb
transformers
torch
sentence-transformers
huggingface-hub
